+++ NEWS.md
@@ -60,6 +60,10 @@
 
 * New FAQ page and updated installation instructions for installing Python dependencies (#424)
 
+* New `greta_sitrep()` function to generate a situation report of the software
+  that is available for use, and also initialising python so `greta` is ready to
+  use. (#441)
+
 # greta 0.3.1
 
 This release is predominantly a patch to make greta work with recent versions of TensorFlow and TensorFlow Probability, which were not backward compatible with the versions on which greta previously depended. From this release forward, greta will depend on specific (rather than minimum) versions of these two pieces of software to avoid it breaking if more changes are made to the APIS of these packages.